ไฟล์ firmware/Webserver-Django/requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.3.1
bcrypt==3.1.7
beautifulsoup4==4.9.3
big-O-calculator==0.0.9.8.4
blinker==1.4
Brlapi==0.7.0
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
cloudpickle==1.6.0
cmake==3.18.4.post1
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.5
djangorestframework==3.12.4
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
gym==0.18.0
httplib2==0.14.0
idna==2.8
kazam==1.4.5
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.3.3
monotonic==1.5
mysqlclient==2.0.3
netifaces==0.10.4
numpy==1.19.4
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.46
pandas==1.2.0
paramiko==2.6.0
pbr==5.5.1
pexpect==4.6.0
Pillow==7.2.0
protobuf==3.6.1
pycairo==1.16.2
pycups==1.9.73
pygame==2.0.1
pyglet==1.5.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
PyQt5==5.14.1
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2020.5
pyxdg==0.26
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
scipy==1.5.4
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.21
six==1.14.0
soupsieve==2.1
sqlparse==0.4.1
stevedore==3.3.0
system-service==0.3
systemd-python==234
tabulate==0.8.7
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.4.2
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wadllib==1.3.3
xkit==0.0.0
